code restructure to enable package install
fixes #91

removed keyword to IntSlider that was causing test_pages to fail with Panel v1.18.1

renamed variables in __main__.py, addressed logging FIXME in __main__.py, updated instructions to run in README
