+++ setup.cfg
@@ -1,0 +1,50 @@
+[metadata]
+name = tsqc
+author = Tskit Developers
+author_email = admin@tskit.dev
+license = MIT
+description = Visual QC of ARGs
+project_urls =
+        Bug Tracker = https://github.com/tskit-dev/tsqc/issues
+        Source Code = https://github.com/tskit-dev/tsqc
+classifiers:
+	Intended Audience :: Science/Research
+        License :: OSI Approved :: MIT License
+        Programming Language :: Python :: 3
+        Topic :: Scientific/Engineering
+        Topic :: Scientific/Engineering :: Bio-Informatics
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://tskit.dev/tsqc/
+platforms =
+	POSIX
+	Windows
+	MacOS X
+keywords = 
+	tree sequence
+	ancestral recombination graph
+	tskit
+	quality control
+
+[options]
+packages = tsqc
+python_requires = >=3.8
+install_requires = 
+	click
+	daiquiri
+	panel
+	diskcache
+	hvplot
+	xarray
+	datashader
+	tskit
+	seaborn
+	pytest
+	tszip
+	appdirs
+setup_requires =
+	setuptools
+
+[tool:pytest]
+testpaths = 
+	tests
