+++ tsqc/config.py
