+++ tsqc/pages/__init__.py
@@ -1,0 +1,7 @@
+from . import edge_explorer  # noqa: F401
+from . import edges  # noqa: F401
+from . import mutations  # noqa: F401
+from . import nodes  # noqa: F401
+from . import overview  # noqa: F401
+from . import popgen  # noqa: F401
+from . import trees  # noqa: F401
